Moved network related headers to torrent/net.
