+++ src/torrent/net/socket_address_key.cc
@@ -1,0 +1,5 @@
+// Copyright (C) 2005-2014, Jari Sundell
+// All rights reserved.
+
+#include "config.h"
+
